requirements.txt
fastapi[standard]
requests
easyocr
google-generativeai
google-cloud-translate
opencv-python
numpy
Pillow
python-dotenv
transformers
addict
matplotlib
easydict 
einops
dotenv
